requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.6
django-allauth==0.44.0
django-autoslug==1.9.8
django-crispy-forms==1.11.0
django-email-verification==0.1.0
django-etc==1.3.0
django-hitcount==1.3.3
django-recaptcha==2.0.6
django-taggit==1.3.0
django-tinymce==3.2.0
Django-Verify-Email==0.0.5
idna==2.10
oauthlib==3.1.0
Pillow==8.1.0
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
